refactor: sever bench-matrix and lab-workspace-provenance agent edges
Two Stage-3 prep severances so the last core files that stay in core stop
importing agent-task types before the homeboy-agents git-mv:

1. bench cross-rig comparison built an agent-task matrix (plan + aggregate)
   from rig entries using agent-task types. Inverted behind a
   BenchAgentTaskMatrixProvider hook: the bench comparison report's
   agent_task_plan/agent_task_aggregate fields become opaque JSON; the
   matrix-building logic moves to the agent-side provider (which takes core's
   pub RigBenchEntry and returns the matrix as JSON).

2. extension/trace/canonicality called agent_task_scheduler::
   lab_workspace_provenance (a provider registry with zero agent-type
   dependencies, misfiled under agent_task_scheduler). Relocated it to
   crate::lab_workspace_provenance (core proper) and repointed all consumers
   (core, cli, lab-runner).

With no provider registered the bench matrix is simply omitted. Behavior
preserved; the bench matrix integration test registers the provider and
deserializes the opaque fields.

Completes the Stage-3 edge-severance prep: core is now agent-task-type-free
except for the modules that move with agent + test-only edges.
