+++ crates/homeboy-core/src/lib.rs
@@ -147,6 +147,7 @@ pub mod io;
 pub mod keychain;
 pub mod lab_offload;
 pub mod lab_routing;
+pub mod lab_workspace_provenance;
 pub use homeboy_lifecycle_contract::lifecycle;
 pub mod loop_lifecycle;
 pub mod markdown;
