+++ tests/core/extension/bench/report_comparison_test.rs
@@ -483,6 +483,9 @@ fn no_axis_multi_rig_comparison_omits_axis_diffs() {
 
 #[test]
 fn axis_diffs_cover_two_by_two_rig_matrix() {
+    // The bench comparison builds its agent-task matrix through a provider hook;
+    // register the agent-task implementation so the matrix is populated.
+    crate::agent_task::bench_matrix_provider::register();
     let entries = vec![
         entry(
             "studio-sdk-standard",
@@ -543,7 +546,9 @@ fn axis_diffs_cover_two_by_two_rig_matrix() {
     let (out, _) = aggregate_comparison_with_axes("studio".into(), 10, entries, &axes_by_rig);
 
     assert_eq!(out.axis_diffs.len(), 4);
-    let agent_task_plan = out.agent_task_plan.as_ref().expect("agent task plan");
+    let agent_task_plan: crate::agent_task::AgentTaskMatrixPlan =
+        serde_json::from_value(out.agent_task_plan.clone().expect("agent task plan"))
+            .expect("agent task plan deserializes");
     assert_eq!(agent_task_plan.plan_id, "bench/studio");
     assert_eq!(agent_task_plan.cells.len(), 4);
     let sdk_standard_cell = agent_task_plan
@@ -559,10 +564,12 @@ fn axis_diffs_cover_two_by_two_rig_matrix() {
         Some("bench/studio")
     );
     assert_eq!(sdk_standard_cell.task.metadata["matrix"]["runtime"], "sdk");
-    let agent_task_aggregate = out
-        .agent_task_aggregate
-        .as_ref()
-        .expect("agent task aggregate");
+    let agent_task_aggregate: crate::agent_task::AgentTaskMatrixAggregate = serde_json::from_value(
+        out.agent_task_aggregate
+            .clone()
+            .expect("agent task aggregate"),
+    )
+    .expect("agent task aggregate deserializes");
     assert!(agent_task_aggregate.passed);
     assert_eq!(agent_task_aggregate.cells.len(), 4);
     assert_eq!(
